extract useAutoCompleteLeafs hook
