+++ packages/graphiql-react/src/editor/context.tsx
@@ -1,8 +1,6 @@
-import { fillLeafs, GetDefaultFieldNamesFn } from '@graphiql/toolkit';
 import { DocumentNode, OperationDefinitionNode } from 'graphql';
 import { VariableToType } from 'graphql-language-service';
-import { ReactNode, useCallback, useMemo, useState } from 'react';
-import { useSchemaContext } from '../schema';
+import { ReactNode, useMemo, useState } from 'react';
 
 import { createContextHook, createNullableContext } from '../utility/context';
 import { CodeMirrorEditor } from './types';
@@ -15,7 +13,6 @@ export type CodeMirrorEditorWithOperationFacts = CodeMirrorEditor & {
 };
 
 export type EditorContextType = {
-  autoCompleteLeafs(): string | undefined;
   headerEditor: CodeMirrorEditor | null;
   queryEditor: CodeMirrorEditorWithOperationFacts | null;
   responseEditor: CodeMirrorEditor | null;
@@ -30,16 +27,7 @@ export const EditorContext = createNullableContext<EditorContextType>(
   'EditorContext',
 );
 
-type EditorContextProviderProps = {
-  children: ReactNode;
-  getDefaultFieldNames?: GetDefaultFieldNamesFn;
-};
-
-export function EditorContextProvider(props: EditorContextProviderProps) {
-  const { schema } = useSchemaContext({
-    nonNull: true,
-    caller: EditorContextProvider,
-  });
+export function EditorContextProvider(props: { children: ReactNode }) {
   const [headerEditor, setHeaderEditor] = useState<CodeMirrorEditor | null>(
     null,
   );
@@ -54,53 +42,8 @@ export function EditorContextProvider(props: EditorContextProviderProps) {
     null,
   );
 
-  const autoCompleteLeafs = useCallback<
-    EditorContextType['autoCompleteLeafs']
-  >(() => {
-    if (!queryEditor) {
-      return;
-    }
-
-    const query = queryEditor.getValue();
-    const { insertions, result } = fillLeafs(
-      schema,
-      query,
-      props.getDefaultFieldNames,
-    );
-    if (insertions && insertions.length > 0) {
-      queryEditor.operation(() => {
-        const cursor = queryEditor.getCursor();
-        const cursorIndex = queryEditor.indexFromPos(cursor);
-        queryEditor.setValue(result || '');
-        let added = 0;
-        const markers = insertions.map(({ index, string }) =>
-          queryEditor.markText(
-            queryEditor.posFromIndex(index + added),
-            queryEditor.posFromIndex(index + (added += string.length)),
-            {
-              className: 'autoInsertedLeaf',
-              clearOnEnter: true,
-              title: 'Automatically added leaf fields',
-            },
-          ),
-        );
-        setTimeout(() => markers.forEach(marker => marker.clear()), 7000);
-        let newCursorIndex = cursorIndex;
-        insertions.forEach(({ index, string }) => {
-          if (index < cursorIndex) {
-            newCursorIndex += string.length;
-          }
-        });
-        queryEditor.setCursor(queryEditor.posFromIndex(newCursorIndex));
-      });
-    }
-
-    return result;
-  }, [props.getDefaultFieldNames, queryEditor, schema]);
-
   const value = useMemo<EditorContextType>(
     () => ({
-      autoCompleteLeafs,
       headerEditor,
       queryEditor,
       responseEditor,
@@ -110,13 +53,7 @@ export function EditorContextProvider(props: EditorContextProviderProps) {
       setResponseEditor,
       setVariableEditor,
     }),
-    [
-      autoCompleteLeafs,
-      headerEditor,
-      queryEditor,
-      responseEditor,
-      variableEditor,
-    ],
+    [headerEditor, queryEditor, responseEditor, variableEditor],
   );
 
   return (
