+++ packages/graphiql-react/src/editor/index.ts
@@ -6,7 +6,12 @@ import {
   useEditorContext,
 } from './context';
 import { useHeaderEditor } from './header-editor';
-import { useCopyQuery, useMergeQuery, usePrettifyEditors } from './hooks';
+import {
+  useAutoCompleteLeafs,
+  useCopyQuery,
+  useMergeQuery,
+  usePrettifyEditors,
+} from './hooks';
 import { useQueryEditor } from './query-editor';
 import { useResponseEditor } from './response-editor';
 import { useVariableEditor } from './variable-editor';
@@ -25,6 +30,7 @@ export {
   ImagePreview,
   EditorContext,
   EditorContextProvider,
+  useAutoCompleteLeafs,
   useCopyQuery,
   useEditorContext,
   useHeaderEditor,
