+++ packages/graphiql-react/src/index.ts
@@ -3,6 +3,7 @@ import {
   EditorContextProvider,
   ImagePreview,
   onHasCompletion,
+  useAutoCompleteLeafs,
   useCopyQuery,
   useEditorContext,
   useHeaderEditor,
@@ -57,6 +58,7 @@ export {
   EditorContextProvider,
   ImagePreview,
   onHasCompletion,
+  useAutoCompleteLeafs,
   useCopyQuery,
   useEditorContext,
   useHeaderEditor,
