+++ packages/graphiql/__mocks__/@graphiql/react.ts
@@ -11,6 +11,7 @@ import {
   SchemaContextProvider,
   StorageContext,
   StorageContextProvider,
+  useAutoCompleteLeafs,
   useCopyQuery,
   useEditorContext,
   useExplorerContext,
@@ -54,6 +55,7 @@ export {
   SchemaContextProvider,
   StorageContext,
   StorageContextProvider,
+  useAutoCompleteLeafs,
   useCopyQuery,
   useEditorContext,
   useExplorerContext,
